More code from 2008 - I think this was the first update
This code is from Wed, Aug 13, 2008 and contains code which I believe
went into the first app update that I released to the App Store. It
added a global high scores list, displayed in a web view, backed by a
PHP script that I ran on a shared web host.